Update patch set 1
Patch Set 1: Code-Review-1

The description is linked to the VoLTE use-case and it not generic enough to be included in the Modeling project. How this modeling is aligned with AAI and SDC modeling concepts ?
There is some inconsistency: CP are introduced in the abbreviation while you propose to terminate the VPN by 2 sites. One site is only connected to 1 network according to the text. 
Such use-case based description should be introduced in the Use-Case documentation part.

Patch-set: 1
Reviewer: Gerrit User 269 <269@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1