Update patch set 1
Patch Set 1: Code-Review+1

+1 to Eric's suggestion on documenting use-case template/artifacts. We could do that since we will have them and share them anyway. 

But to clarify, this document is not specific to VoLTE, but represents one of the key features for SDC and SDNC in DCI orchestration and configuration. It is a different perspective from the ETSI NFV deployment model. I believe we have statement in the document that the tunnel is modeled as a special VL resource. 
But above all, it is meant for documenting what is ONAP R1 is capable of in orchestrating L2/L3 VPN and help people leverage it to build their own use case in addition to the two examples we are demonstrating as part of the release. 
I would be happy to revise it later to clearly state its relationship with ETSI NFV VL and VoLTE or CPE usecase, and provide input to VoLTE artifacts if needed.

Patch-set: 1
Reviewer: Gerrit User 2123 <2123@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1