Update to Common
Finalized a draft root proposal and created a shell location package
with a place class referenced in the root proposal.  There are a number
of 'temporary' duplicate classes in the root proposal intentionally left
in the diagrams package.   These will be deleted and replaced with the
correct classes from various submodels when appropriate.

Issue-ID: MODELING-104

Change-Id: If837a150c2a13fcae6742cae1228238e19241243
Signed-off-by: Kevin Scaggs <ks0567@att.com>