Update patch set 1
Patch Set 1: Code-Review+1

It seems to me that you want to model the situation where a business interaction could reference multiple other business interactions, or in other words, a nesting relationship. I'm a little curious about why not modeling it as a "reference" relationship starting from the "BusinessInteraction" class and pointing to itself. Actually I'm not so clear about what's the purpose and usage for the "BusinessInteractionRelationship" class, and why it's the endpoint of the "reference" and "referenceBy" associations. Another question is about whether these associations could be inherited to all the subclasses. The same questions apply to the "BusinessInteractionItem" as well...

Patch-set: 1
Reviewer: Gerrit User 3759 <3759@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1