Update patch set 1
Patch Set 1: Code-Review+1

A comment on the usage of multiplicity and default value: though I understand the intention is to show the possible values in the diagram, I'm afraid it's not the correct way to use those properties. IMO, the multiplicity represents the possible number of entries an attribute could have, for example, if the multiplicity of the attribute "color" is 3, it should mean that the value of "color" would always be a tuple of 3 strings; and the default value should represent the value automatically assigned if no inputs are given, but not list all the possible values. In a word, I suggest to use another way to represent the value range.

Patch-set: 1
Reviewer: Gerrit User 3759 <3759@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1