Updates to Common
As part of updates to allotted resource, was reminded that the resource
objects were not rooted (this was approved with temporary objects).
Decided to make cross model updates to assure resource, service, nf,
pnf, vnf, ... are all appropriately rooted as approved.   Some were
rooted, some were not.    Also deleted all the temporary objects from
common, which is also helpful from a glossary point of view.

Issue-ID: MODELING-104
Change-Id: I891942220b44bce19bab918958691d3605418978
Signed-off-by: Kevin Scaggs <ks0567@att.com>