Many changes to bring NSD into alignment with ETSI IFA014 v2.7.1 - Support for Sapd - Support for nested NSD - Support for vnffgd - Updated NsDf - Updated Nslifecyclemanagementscripts
Added a optional reference to an ETSI NSD in the ServiceDescriptor
Since the ServiceDescriptor may be nested an NSD may be referenced by a nested ServiceDescriptor.
Therefore, several NSD may be referenced by a hierarchy of Service Descriptors.
Resources defined by a an NSD will not be described by the ServiceDescriptor

Issue-ID: MODELING-334

Signed-off-by: Fernando Oliveira <fernando.oliveira@verizon.com>
Change-Id: I0aff2b29c65502dde98c655dbb344864a85628af